% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_tables_title}
\alias{all_tables_title}
\title{Title of tables}
\usage{
all_tables_title()
}
\value{
A tibble of title of all tables
}
\description{
\code{all_tables_title()} returns title of all tables. It is helpful to get familiar with available tables. Each table contains
a number of data set.
}
\examples{
all_tables_title()
}
\seealso{
\url{https://www.federalreserve.gov/apps/fof/FOFTables.aspx}
}
