% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_series}
\alias{get_series}
\title{Data}
\usage{
get_series(choice = c("series", "tables"))
}
\arguments{
\item{choice}{can be "series" or "tables}
}
\value{
series or tables
}
\description{
The function returns data series or tables. Due to CRAN package size requirements, the data is stored on github.
}
\details{
\code{get_series()} load entire series or tables. For example: get_series(choice="series")
\code{series_data} contains all series available in flow of funds and \code{tables_data} contains all tables and series
}
\section{series_data}{

\itemize{
\item date - Time-period
\item value - Observed-Value
\item SERIES_NAME - name of a series. There are more than twenty thousands series exist
\item CURRENCY - USD or NA. NA means series is percent or index
\item SERIES_PREFIX - Each prefix letters indicate the series type. see fedz1:::meaning_of_prefix()
\item UNIT - Currency, percent or index
\item UNIT_MULT - 1,1,000 or 1,000,000
\item short description _ description of the series
\item FREQ - A:annually data; Q:quarterly data
}
}

\section{tables_data}{

\itemize{
\item tname -  name of a table
\item SERIES_NAME - name of a series. There are more than twenty thousands series exist
\item Description - Long Description of series
\item date - Time-period
\item value - Observed-Value
\item UNIT - Currency, percent or index
}
}

\seealso{
\url{https://www.federalreserve.gov/datadownload/Choose.aspx?rel=z1}

\url{https://www.federalreserve.gov/apps/fof/FOFTables.aspx}
}
