% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meaning_of_prefix}
\alias{meaning_of_prefix}
\title{Definition of series prefix}
\usage{
meaning_of_prefix(prefix)
}
\arguments{
\item{prefix}{is SERIES_PREFIX. Value could be FA, FC, FG, FI, FL, FR, FS, FU, FV, LA, LM, PC.}
}
\value{
character for definition of series
}
\description{
\code{meaning_of_prefix()} returns definition of series. Each series start with two letters. Each prefix letters indicate the series type.
}
\examples{
meaning_of_prefix("FL")
}
\seealso{
\url{https://www.federalreserve.gov/apps/fof/SeriesStructure.aspx}
}
