% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{mprefix}
\alias{mprefix}
\title{Meaning of prefix}
\format{mprefix

A data frame with 12 rows and 2 columns:
\describe{
\item{SERIES_PREFIX}{prefix of series}
\item{meaning }{meaning of a prefix}
}}
\source{
\url{https://www.federalreserve.gov/apps/fof/SeriesStructure.aspx}
}
\usage{
mprefix
}
\description{
Each prefix letters indicate the series type
}
\keyword{datasets}
