% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{prds}
\alias{prds}
\title{Title of series with a short description}
\format{prds

A data frame with 11,291 rows and 2 columns:
\describe{
\item{title1}{series name}
\item{short description}{description of the series}
}}
\source{
\url{https://www.federalreserve.gov/datadownload/Choose.aspx?rel=z1}
}
\usage{
prds
}
\description{
Data is obtained from a web scraping from flow of funds website
}
\keyword{datasets}
