% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_series}
\alias{search_series}
\title{search series}
\usage{
search_series(series)
}
\arguments{
\item{series}{is SERIES name}
}
\value{
A tibble of series
}
\description{
\code{search_series()} return all series with the search. For example: search_series("Real estate investment trusts")
}
