% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{search_tables}
\alias{search_tables}
\title{search tables}
\usage{
search_tables(tables)
}
\arguments{
\item{tables}{is table name}
}
\value{
A tibble of series in a table
}
\description{
\code{search_tables()} return series associated with a table
}
\examples{
search_tables("Nonfinancial Corporate Business")
}
