% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{table_descr}
\alias{table_descr}
\title{Description of a table}
\usage{
table_descr(table_title)
}
\arguments{
\item{table_title}{is name of a single table or vector of multiple tables. You can obtain title of tables by \code{all_tables_title()}}
}
\value{
character for definition of a table(s)
}
\description{
\code{table_descr()} return the definition of a table. Each table contains a number of data set.
}
\examples{
table_descr('Credit Unions')
table_descr(c('Net Capital Transfers',"Closed-End Funds"))
}
\seealso{
\url{https://www.federalreserve.gov/apps/fof/Guide/z1_tables_description.pdf}
}
