% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{table_detail}
\alias{table_detail}
\title{Title of tables}
\format{table_detail

A data frame with 198 rows and 3 columns:
\describe{
\item{summary}{title as shows on the source website}
\item{name}{title of the table}
\item{code}{code of each table as shows on the source website}
}}
\source{
\url{https://www.federalreserve.gov/apps/fof/FOFTables.aspx}
}
\usage{
table_detail
}
\description{
Data is obtained from a web scraping from Financial Accounts Guide
}
\keyword{datasets}
